+++ blazorbootstrap/Components/Form/DragAndDropFileInput/DragAndDropFileInput.razor.cs
@@ -1,0 +1,48 @@
+using Microsoft.AspNetCore.Components.Web;
+
+namespace BlazorBootstrap;
+
+public partial class DragAndDropFileInput : FileInputBase
+{
+    #region Fields and Constants
+
+    private int dragDepth;
+
+    #endregion
+
+    #region Methods
+
+    #region Protected Methods
+
+    protected void OnDragEnter(DragEventArgs _)
+    {
+        dragDepth++;
+        StateHasChanged();
+    }
+
+    protected void OnDragLeave(DragEventArgs _)
+    {
+        dragDepth = Math.Max(0, dragDepth - 1);
+        StateHasChanged();
+    }
+
+    protected void OnDrop(DragEventArgs _)
+    {
+        dragDepth = 0;
+        StateHasChanged();
+    }
+
+    #endregion
+
+    #endregion
+
+    #region Properties, Indexers
+
+    protected override string? ClassNames =>
+        BuildClassNames(Class,
+            ("bb-drag-and-drop-file-input", true),
+            (BackgroundColor.ToBackgroundClass(), BackgroundColor != BackgroundColor.None),
+            ("bb-drag-and-drop-file-input--dragging", dragDepth > 0));
+
+    #endregion
+}
