+++ blazorbootstrap/Components/Form/DragAndDropFileInput/DragAndDropFileInput.razor.css
@@ -1,0 +1,62 @@
+.bb-drag-and-drop-file-input {
+    position: relative;
+    min-height: 9rem;
+    overflow: hidden;
+    border: 2px dashed var(--bs-border-color);
+    border-radius: var(--bs-border-radius);
+}
+
+.bb-drag-and-drop-file-input--dragging {
+    border-color: var(--bs-primary);
+    background-color: var(--bs-primary-bg-subtle);
+}
+
+.bb-drag-and-drop-file-input__native-input {
+    position: absolute;
+    inset: 0;
+    z-index: 1;
+    width: 100%;
+    height: 100%;
+    cursor: pointer;
+    opacity: 0;
+}
+
+.bb-drag-and-drop-file-input__content {
+    display: flex;
+    min-height: 9rem;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: .25rem;
+    padding: 1rem;
+    text-align: center;
+    pointer-events: none;
+}
+
+.bb-drag-and-drop-file-input__title {
+    font-weight: 500;
+}
+
+.bb-drag-and-drop-file-input__hint {
+    color: var(--bs-secondary-color);
+    font-size: .875em;
+}
+
+.bb-file-input__files {
+    list-style: none;
+    margin: .75rem 0 0;
+    padding: 0;
+}
+
+.bb-file-input__file {
+    align-items: center;
+    display: flex;
+    gap: .75rem;
+    justify-content: space-between;
+    padding: .25rem 0;
+}
+
+.bb-file-input__errors {
+    color: var(--bs-danger-text-emphasis);
+    margin-top: .75rem;
+}
