+++ blazorbootstrap/Components/Form/FileInput/FileInput.razor
@@ -1,29 +1,14 @@
 @namespace BlazorBootstrap
-@inherits BlazorBootstrapComponentBase
+@inherits FileInputBase
 
-<div @ref="@Element"
-     id="@Id"
-     class="@ClassNames"
-     style="@StyleNames"
-     @attributes="@AdditionalAttributes">
-    <div class="@DropZoneClassNames"
-         @ondragenter="OnDragEnter"
-         @ondragleave="OnDragLeave"
-         @ondragover:preventDefault="true"
-         @ondrop="OnDrop">
-        <InputFile class="bb-file-input__native"
-                   OnChange="OnInputFileChangedAsync"
-                   accept="@Accept"
-                   multiple="@Multiple"
-                   aria-label="@AriaLabel" />
-        <div class="bb-file-input__content" aria-hidden="true">
-            <span class="bb-file-input__title">Drop files here or click to browse</span>
-            @if (Multiple)
-            {
-                <span class="bb-file-input__description">Multiple files are allowed.</span>
-            }
-        </div>
-    </div>
+<div @ref="@Element" @attributes="@AdditionalAttributes">
+    <InputFile id="@Id"
+               class="@ClassNames"
+               style="@StyleNames"
+               OnChange="OnInputFileChangedAsync"
+               accept="@Accept"
+               multiple="@Multiple"
+               aria-label="@AriaLabel" />
 
     @if (files.Any())
     {
@@ -32,10 +17,7 @@
             {
                 <li class="bb-file-input__file">
                     <span>@file.Name</span>
-                    <button type="button"
-                            class="btn-close"
-                            aria-label="Remove @file.Name"
-                            @onclick="async () => await RemoveFileAsync(file)"></button>
+                    <button type="button" class="btn-close" aria-label="Remove @file.Name" @onclick="async () => await RemoveFileAsync(file)"></button>
                 </li>
             }
         </ul>
